+++ crates/goose/src/agents/prompt_manager.rs
@@ -3,7 +3,7 @@ use serde_json::Value;
 use std::collections::HashMap;
 
 use crate::agents::extension::ExtensionInfo;
-use crate::providers::get_current_model;
+use crate::providers::base::get_current_model;
 use crate::{config::Config, prompt_template};
 
 pub struct PromptManager {
@@ -84,8 +84,6 @@ impl PromptManager {
         // First check the global store, and only if it's not available, fall back to the provided model_name
         let model_to_use: Option<String> =
             get_current_model().or_else(|| model_name.map(|s| s.to_string()));
-
-        print!("Model to use: {:?}", model_to_use);
 
         // Conditionally load the override prompt or the global system prompt
         let base_prompt = if let Some(override_prompt) = &self.system_prompt_override {
